// Neelife/src/app/game/shared/interfaces/personnage.ts
export interface IPersonnage {

  id: number;
  id_utilisateur: number;
  nom: string;
  prenom: string;
  situationPersonnage: any;
  interactionsPersonnage: any;
  diplomesPersonnages: any;
  santePersonnage: any;
  id_ville: any;
  entreprises: any;
  compteBancaires: any;
  addictionsPersonnages: any;


}